feat(investors): unify personalized portal narrative
